fix(warnings): eliminate all compiler warnings
- Keyboard.cpp: add default member initializers to MAPPED_KEY struct
  (fixes ~500x missing-field-initializers warnings)
- CMakeLists.txt: move -Wno-implicit-function-declaration to C-only flag
  (fixes 63x 'valid for C/ObjC but not C++' warnings)
- HardDisk.cpp: zero-initialize RS_IDE::abReserved with {}
- Debug.cpp: cast enum color constants to uint8_t in ternaries,
  remove always-true unsigned >= 0 comparison
- GUI.cpp, GUIDlg.cpp: cast enum/int mismatches in ternary expressions
- sp0256.cpp: use ~0u instead of -1/~0 to avoid signed left-shift UB
- CMakeLists.txt: suppress warnings in vendored libspectrum C files

Build now produces zero compiler warnings. Closes SamCoupe-2g3.

+++ components/simcoupe_core/Base/GUIDlg.cpp
@@ -1264,7 +1264,7 @@ void ImportDialog::OnNotify(Window* pWindow_, int nParam_)
             return;
         }
 
-        unsigned int uPage = (s_uAddr < 0x4000) ? ROM0 : s_uPage;
+        unsigned int uPage = (s_uAddr < 0x4000) ? unsigned(ROM0) : s_uPage;
         unsigned int uOffset = s_uOffset, uLen = 0x400000;  // 4MB max import
         size_t uRead = 0;
 
@@ -1331,7 +1331,7 @@ void ExportDialog::OnNotify(Window* pWindow_, int nParam_)
             return;
         }
 
-        unsigned int uPage = (s_uAddr < 0x4000) ? ROM0 : s_uPage;
+        unsigned int uPage = (s_uAddr < 0x4000) ? unsigned(ROM0) : s_uPage;
         unsigned int uOffset = s_uOffset, uLen = s_uLength;
         size_t uWritten = 0;
